moved models to db folder
