+++ server/auth/verifyToken.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const User = require('../models/User');
+const User = require('../db/models/User');
 
 // Middleware that checks if a user has a token
 // returns the user's data if true
